Parsing response and rendering list.

+++ app/src/main/java/apps/iloyarte/melichallenge/api/SearchResponse.kt
@@ -122,7 +122,7 @@ data class Result(
         val attribute_group_name: String,
         val id: String,
         val name: String,
-        val source: Int,
+        val source: Double,
         val value_id: String,
         val value_name: String,
         val value_struct: Any
